Move fromJWK and Raw tests to multikey.
