+++ test/assertions.js
@@ -225,6 +225,96 @@ export function testFrom({serializedKeyPair, id, keyType}) {
   });
 }
 
+export function testJWK({curve}) {
+  it('should round-trip secret JWKs', async () => {
+    const keyPair = await EcdsaMultikey.generate({
+      id: '4e0db4260c87cc200df3',
+      curve
+    });
+    const jwk1 = await EcdsaMultikey.toJwk({keyPair, secretKey: true});
+    expect(jwk1.d).to.exist;
+    const keyPairImported = await EcdsaMultikey.fromJwk(
+      {jwk: jwk1, secretKey: true});
+    const jwk2 = await EcdsaMultikey.toJwk(
+      {keyPair: keyPairImported, secretKey: true});
+    expect(jwk1).to.eql(jwk2);
+  });
+
+  it('should round-trip public JWKs', async () => {
+    const keyPair = await EcdsaMultikey.generate({
+      id: '4e0db4260c87cc200df3',
+      curve
+    });
+    const jwk1 = await EcdsaMultikey.toJwk({keyPair});
+    expect(jwk1.d).to.not.exist
+    const keyPairImported = await EcdsaMultikey.fromJwk({jwk: jwk1});
+    const jwk2 = await EcdsaMultikey.toJwk({keyPair: keyPairImported});
+    expect(jwk1).to.eql(jwk2);
+  });
+}
+
+export function testRaw({curve}) {
+  it('should import raw public key', async () => {
+    const keyPair = await EcdsaMultikey.generate({curve});
+
+    // first export
+    const expectedPublicKey = base58.decode(
+      keyPair.publicKeyMultibase.slice(1)).slice(2);
+    const {publicKey} = await keyPair.export({publicKey: true, raw: true});
+    expect(expectedPublicKey).to.deep.equal(publicKey);
+
+    // then import
+    const imported = await EcdsaMultikey.fromRaw({curve, publicKey});
+
+    // then re-export to confirm
+    const {publicKey: publicKey2} = await imported.export(
+      {publicKey: true, raw: true});
+    expect(expectedPublicKey).to.deep.equal(publicKey2);
+  });
+
+  it('should import raw secret key', async () => {
+    const keyPair = await EcdsaMultikey.generate({curve});
+
+    // first export
+    const expectedSecretKey = base58.decode(
+      keyPair.secretKeyMultibase.slice(1)).slice(2);
+    const {secretKey, publicKey} = await keyPair.export(
+      {secretKey: true, raw: true});
+    expect(expectedSecretKey).to.deep.equal(secretKey);
+
+    // then import
+    const imported = await EcdsaMultikey.fromRaw(
+      {curve, secretKey, publicKey});
+
+    // then re-export to confirm
+    const {secretKey: secretKey2} = await imported.export(
+      {secretKey: true, raw: true});
+    expect(expectedSecretKey).to.deep.equal(secretKey2);
+  });
+
+  it('should import raw secret key for key agreement', async () => {
+    const keyPair = await EcdsaMultikey.generate({curve, keyAgreement: true});
+
+    // first export
+    const expectedSecretKey = base58.decode(
+      keyPair.secretKeyMultibase.slice(1)).slice(2);
+    const {secretKey, publicKey} = await keyPair.export(
+      {secretKey: true, raw: true});
+    expect(expectedSecretKey).to.deep.equal(secretKey);
+
+    // then import
+    const imported = await EcdsaMultikey.fromRaw(
+      {curve, secretKey, publicKey, keyAgreement: true});
+    expect(imported.keyAgreement).to.equal(true);
+
+    // then re-export to confirm
+    const {secretKey: secretKey2} = await imported.export(
+      {secretKey: true, raw: true});
+    expect(expectedSecretKey).to.deep.equal(secretKey2);
+  });
+
+}
+
 function _ensurePublicKeyEncoding({keyPair, publicKeyMultibase, keyType}) {
   keyPair.publicKeyMultibase.startsWith('z').should.be.true;
   publicKeyMultibase.startsWith('z').should.be.true;
